feat(RestApi): move body parsing to schemaBodyJson
